rest/requirements.txt
Flask
flask-accept
# Flask-Migrate
# Flask-SQLAlchemy
MarkupSafe
# psycopg2-binary
pyyaml
requests
# SQLAlchemy
GitPython
